App-wide global code rebase
- Introduced the `global_schema` file that handles all app-wide global variable assignments and storage

- Splitted screen, frame, and dialog classes in `main.py` to individual class files under `handler` folder
