+++ src/simon_petrus/handler/frame/__init__.py
@@ -1,0 +1,6 @@
+"""
+Simon Petrus
+AGPL-3.0-licensed
+Copyright (C) GKI Salatiga 2024
+Written by Samarthya Lykamanuella (github.com/groaking)
+"""
