+++ src/simon_petrus/refs/__init__.py
@@ -1,0 +1,47 @@
+"""
+Simon Petrus
+AGPL-3.0-licensed
+Copyright (C) GKI Salatiga 2024
+Written by Samarthya Lykamanuella (github.com/groaking)
+
+This module handles the global external (internet) references along with their links
+that are used in this entire package.
+
+---
+REFERENCES:
+    [1] Dynamic loading of .ui files in Python
+    - https://github.com/eyllanesc/stackoverflow/tree/master/questions/53899209
+    - https://stackoverflow.com/a/53926554
+    [2] "on_..._clicked()" functions are automatically connected to the respective slot
+    - https://forum.qt.io/post/522391
+    [3] Use "@pyqtSlot()" to prevent the slot from firing twice
+    - https://stackoverflow.com/a/77741959
+    [4] Remove all child items from a layout
+    - https://stackoverflow.com/a/9375273
+    - https://stackoverflow.com/a/9383780
+    [5] Prevents GUI freezing
+    - https://www.xingyulei.com/post/qt-threading/index.html
+    [6] AES encryption of strings
+    - https://onboardbase.com/blog/aes-encryption-decryption
+    - https://github.com/Legrandin/pycryptodome
+    [7] WordPress REST API references
+    - https://medium.com/@vicgupta/how-to-upload-a-post-or-image-to-wordpress-using-rest-api-220fe046ff7a
+    [8] Playing GIF images in PyQt5
+    - https://stackoverflow.com/q/10261265
+    - https://www.perplexity.ai/search/qtdesigner-python-displaying-g-cPTPIewkRHS1213kS0t2nQ
+    [9] PyQt5 multithreading to prevent screen freeze
+    - https://stackoverflow.com/a/65447493
+    - https://github.com/shailshouryya/save-thread-result
+    [10] Prevent window resizing in PyQt5
+    - https://stackoverflow.com/a/13775478
+    [11] Capitalize the first letter of words in a sentence
+    - https://stackoverflow.com/a/1549644
+    [12] Displaying image in PyQt5
+    - https://stackoverflow.com/a/51431109
+    [13] Sorting a QListWidget element.
+    - https://www.geeksforgeeks.org/pyqt5-qlistwidget-making-sorting-enabled/
+    [14] Detecting the mime type of a file
+    - https://stackoverflow.com/a/2753385
+    [15] Defining global variables across files and modules.
+    - https://stackoverflow.com/a/13034908
+"""
